+++ event-calendar/app/routes/admin.categories/$categoryId.delete.tsx
@@ -11,8 +11,6 @@ import {
   Grid,
   GridItem,
   Text,
-  // FormErrorMessage,
-  // FormHelperText,
 } from "@chakra-ui/react";
 
 interface CategoryDeleteModalProps {
